commit 适配主题颜色，修复打不开简书等地址连接，添加说明

+++ app/src/main/java/com/kotlin/mvvm/ui/integral/MyIntegralActivity.kt
@@ -42,7 +42,7 @@ class MyIntegralActivity : BaseActivity() {
         }
         mViewModel.mUserIntegralBean.observe(this){
             headerView.tvTotalIntegral.text = it.coinCount.toString()
-            headerView.tvLevelRank.text = getString(R.string.level_rank, it.level, it.rank.toInt())
+            headerView.tvLevelRank.text = getString(R.string.level_rank, it.level, it.rank)
         }
         mViewModel.mIntegralBean.observe(this) {
             if (it.datas.isNotEmpty()) {
